src/python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
ollama==0.1.7
markdown==3.5.1
beautifulsoup4==4.12.2
numpy==1.24.3
